backend/fastapi/ai/summary/requirements.txt
accelerate==1.5.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bertopic==0.17.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
docopt==0.6.2
executing==2.2.0
filelock==3.18.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.3.6
plotly==6.0.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarg==0.1.10
